minecrafthub/requirements.txt
arrow==1.2.1
asgiref==3.4.1
binaryornot==0.4.4
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cookiecutter==1.7.3
crispy-tailwind==0.5.0
Deprecated==1.2.13
Django==4.0.1
django-admin-interface==0.18.4
django-appconf==1.0.5
django-cleanup==5.2.0
django-colorfield==0.6.3
django-crispy-forms==1.13.0
django-embed-video==1.4.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-maintenancemode==0.11.7
django-redis==5.2.0
django-social-share==2.2.1
django-tailwind==3.0.1
django-tinymce==3.4.0
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
IPy==1.1
Jinja2==3.0.3
jinja2-time==0.2.0
MarkupSafe==2.0.1
packaging==21.3
Pillow==9.0.0
poyo==0.5.0
protobuf==3.19.3
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-slugify==5.0.2
redis==4.1.2
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
